requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
pydantic>=2.7,<3.0
python-dotenv>=1.0,<2.0
tenacity>=8.4,<9.0
lxml>=5.2,<6.0
numpy>=2.0,<3.0
pytest>=8.2,<9.0
